feat: split time and date fields to control composables
