+++ packages/core/src/useDateTime/index.ts
@@ -1,4 +1,6 @@
+export * from './useDateControl';
 export * from './useDateField';
+export * from './useTimeControl';
 export * from './useTimeField';
 export * from './useDateTimeSegment';
 export * from './types';
