+++ packages/core/src/useDateTime/useDateField.ts
@@ -1,110 +1,14 @@
-import { Maybe, Reactivify, StandardSchema } from '../types';
-import type { CalendarProps } from '../useCalendar';
-import { normalizeProps, useUniqId, useCaptureProps } from '../utils/common';
-import { computed, shallowRef, toValue } from 'vue';
-import { exposeField, useFormField } from '../useFormField';
-import { useDateTimeSegmentGroup } from './useDateTimeSegmentGroup';
-import { FieldTypePrefixes } from '../constants';
-import { useDateFormatter, useLocale } from '../i18n';
-import { useErrorMessage, useLabel, useDescription } from '../a11y';
-import { fromDateToCalendarZonedDateTime, useTemporalStore } from './useTemporalStore';
-import { ZonedDateTime, Calendar } from '@internationalized/date';
-import { useInputValidity } from '../validation';
+import { Maybe, Reactivify } from '../types';
+import { normalizeProps } from '../utils/common';
+import { toValue } from 'vue';
+import { exposeField, FieldBaseProps, useFormField } from '../useFormField';
 import { registerField } from '@formwerk/devtools';
-import { useConstraintsValidator } from '../validation/useConstraintsValidator';
+import { DateControlProps, useDateControl } from './useDateControl';
 
-export interface DateFieldProps {
-  /**
-   * The label to use for the field.
-   */
-  label: string;
-
-  /**
-   * The name to use for the field.
-   */
-  name?: string;
-
-  /**
-   * Whether the field is required.
-   */
-  required?: boolean;
-
-  /**
-   * The locale to use for the field.
-   */
-  locale?: string;
-
-  /**
-   * The calendar type to use for the field, e.g. `gregory`, `islamic-umalqura`, etc.
-   */
-  calendar?: Calendar;
-
-  /**
-   * The time zone to use for the field, e.g. `UTC`, `America/New_York`, etc.
-   */
-  timeZone?: string;
-
-  /**
-   * The Intl.DateTimeFormatOptions to use for the field, used to format the date value.
-   */
-  formatOptions?: Intl.DateTimeFormatOptions;
-
-  /**
-   * The description to use for the field.
-   */
-  description?: string;
-
-  /**
-   * The placeholder to use for the field.
-   */
-  placeholder?: string;
-
-  /**
-   * Whether the field is readonly.
-   */
-  readonly?: boolean;
-
-  /**
-   * Whether the field is disabled.
-   */
-  disabled?: boolean;
-
-  /**
-   * The value to use for the field.
-   */
-  value?: Date;
-
-  /**
-   * The model value to use for the field.
-   */
-  modelValue?: Date;
-
-  /**
-   * The schema to use for the field.
-   */
-  schema?: StandardSchema<Date>;
-
-  /**
-   * The minimum date to use for the field.
-   */
-  min?: Date;
-
-  /**
-   * The maximum date to use for the field.
-   */
-  max?: Date;
-}
+export type DateFieldProps = DateControlProps & FieldBaseProps<Date>;
 
 export function useDateField(_props: Reactivify<DateFieldProps, 'schema'>) {
   const props = normalizeProps(_props, ['schema']);
-  const controlEl = shallowRef<HTMLInputElement>();
-  const { locale, direction, timeZone, calendar } = useLocale(props.locale, {
-    calendar: () => toValue(props.calendar),
-    timeZone: () => toValue(props.timeZone),
-  });
-
-  const formatter = useDateFormatter(locale, props.formatOptions);
-  const controlId = useUniqId(FieldTypePrefixes.DateTimeField);
 
   const field = useFormField<Maybe<Date>>({
     label: props.label,
@@ -113,136 +17,15 @@ export function useDateField(_props: Reactivify<DateFieldProps, 'schema'>) {
     disabled: props.disabled,
     initialValue: toValue(props.modelValue) ?? toValue(props.value),
     schema: props.schema,
+    // TODO: Remove once all fields have controls
+    syncModel: false,
   });
 
-  const { element: inputEl } = useConstraintsValidator({
-    type: 'date',
-    required: props.required,
-    value: field.fieldValue,
-    source: controlEl,
-    min: props.min,
-    max: props.max,
-  });
-
-  useInputValidity({ field, inputEl });
-
-  const min = computed(() => fromDateToCalendarZonedDateTime(toValue(props.min), calendar.value, timeZone.value));
-  const max = computed(() => fromDateToCalendarZonedDateTime(toValue(props.max), calendar.value, timeZone.value));
-
-  const temporalValue = useTemporalStore({
-    calendar: calendar,
-    timeZone: timeZone,
-    locale: locale,
-    model: {
-      get: () => field.fieldValue.value,
-      set: value => field.setValue(value),
-    },
-    min,
-    max,
-  });
-
-  function onValueChange(value: ZonedDateTime) {
-    temporalValue.value = value;
-  }
-
-  const { segments } = useDateTimeSegmentGroup({
-    formatter,
-    locale,
-    formatOptions: props.formatOptions,
-    direction,
-    controlEl,
-    temporalValue,
-    readonly: props.readonly,
-    onValueChange,
-    onTouched: () => field.setTouched(true),
-    min,
-    max,
-    dispatchEvent: (type: string) => inputEl.value?.dispatchEvent(new Event(type)),
-  });
-
-  const { labelProps, labelledByProps } = useLabel({
-    for: controlId,
-    label: props.label,
-    targetRef: controlEl,
-  });
-
-  const { descriptionProps, describedByProps } = useDescription({
-    inputId: controlId,
-    description: props.description,
-  });
-
-  const { errorMessageProps, accessibleErrorProps } = useErrorMessage({
-    inputId: controlId,
-    errorMessage: field.errorMessage,
-  });
-
-  const calendarProps = computed(() => {
-    const propsObj: CalendarProps = {
-      label: toValue(props.label),
-      locale: locale.value,
-      name: undefined,
-      calendar: calendar.value,
-      timeZone: timeZone.value,
-      min: toValue(props.min),
-      max: toValue(props.max),
-      field,
-    };
-
-    return propsObj;
-  });
-
-  const controlProps = useCaptureProps(() => {
-    return {
-      id: controlId,
-      role: 'group',
-      ...labelledByProps.value,
-      ...describedByProps.value,
-      ...accessibleErrorProps.value,
-      'aria-disabled': field.isDisabled.value || undefined,
-    };
-  }, controlEl);
+  const control = useDateControl(props, { field });
 
   if (__DEV__) {
     registerField(field, 'Date');
   }
 
-  return exposeField(
-    {
-      /**
-       * The props to use for the control element.
-       */
-      controlProps,
-
-      /**
-       * The props to use for the description element.
-       */
-      descriptionProps,
-
-      /**
-       * The props to use for the label element.
-       */
-      labelProps,
-
-      /**
-       * The props to use for the error message element.
-       */
-      errorMessageProps,
-
-      /**
-       * The datetime segments, you need to render these with the `DateTimeSegment` component.
-       */
-      segments,
-
-      /**
-       * The props to use for the calendar composable/component.
-       */
-      calendarProps,
-
-      /**
-       * The direction of the field.
-       */
-      direction,
-    },
-    field,
-  );
+  return exposeField(control, field);
 }
